reworked directories and file structure
