+++ .gitignore
@@ -57,6 +57,5 @@ typings/
 # dotenv environment variables file
 .env
 
-
 .DS_Store
 .jshintrc
