--- package2.json
@@ -1,17 +1,0 @@
-{
-  "name": "node_demo",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index"
-  },
-  "dependencies": {
-    "cheerio": "^1.0.0-rc.2",
-    "express": "^4.16.2",
-    "pug": "^2.0.0-rc.4"
-  },
-  "author": "",
-  "license": "ISC"
-}
