--- public/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-    <head>
-        <meta charset="utf-8" />
-        <title> Catalysts </title>
-        <link rel="stylesheet" type="text/css" href="styles.css">
-        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-        <script src="./api.js"></script>
-    </head>
-
-    <body>
-
-
-    </body>
-
-</html>
