+++ readme.md
@@ -1,9 +1,9 @@
 # Catalysts for Renewable Energy
 
-### Setup: 
+### Setup:
 * Make sure you're in the project directory (`/CatalystsRE`)
 * Run `npm install` to install node modules
 
-### Run: 
-* Run `npm start` to start the server. 
-* Visit `localhost:3000/graph` in your browser to view the app. 
+### Run:
+* Run `npm start` to start the server.
+* Visit `localhost:3000/graph` in your browser to view the app.
